chore(auth): rename provider level requests.ts to inputs.ts
